backend/requirements.txt
# --- Core API ---
fastapi==0.114.0
uvicorn==0.30.6
joblib>=1.4,<2.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0

# --- Config / Env ---
python-dotenv==1.0.1
pydantic-settings==2.4.0  # nice with Pydantic v2 for typed settings

# --- Auth (Google ID token verification) ---
google-auth==2.33.0

# --- HTTP client (call external nutrition/retailer APIs) ---
httpx==0.27.2

requests>=2.31,<3.0